Public API

TBD

ExprParsers.ExprParsers

ExprParsers.Arg

ExprParsers.Arg_Parsed

ExprParsers.Assignment

ExprParsers.Assignment_Parsed

ExprParsers.Block

ExprParsers.Call

ExprParsers.Call_Parsed

ExprParsers.Expr

ExprParsers.ExprParsed

ExprParsers.ExprParser

ExprParsers.ExprParserWithParsed

ExprParsers.Expr_Parsed

ExprParsers.Function

ExprParsers.Function_Parsed

ExprParsers.Indexed

ExprParsers.Macro

ExprParsers.Macro_Parsed

ExprParsers.Named

ExprParsers.NestedDot

ExprParsers.NestedDot_Parsed

ExprParsers.Reference

ExprParsers.Reference_Parsed

ExprParsers.Signature

ExprParsers.Signature_Parsed

ExprParsers.Type

ExprParsers.TypeAnnotation

ExprParsers.TypeAnnotation_Parsed

ExprParsers.TypeRange

ExprParsers.TypeRange_Parsed

ExprParsers.Type_Parsed

ExprParsers._find_object

ExprParsers._get_struct_name

ExprParsers._parsed_struct_from_parser_struct

ExprParsers.parse_expr

ExprParsers.parse_expr

ExprParsers.to_expr

ExprParsers.@passert

ExprParsers.ExprParsers — Module

ExprParsers

the main interface encompass just three concepts, which seamlessly interact with oneanother

macro @exprparser: easily create definitions for highly flexible and nestable parsers

function parse_expr: compares a matcher with a value, and returns a parsed value

function to_expr: transforms parsed values back to AbstractSyntaxTrees

ExprParsers.Arg — Type

parses:

a
a::B
::B
a = c
a::B = c

ExprParsers.Arg_Parsed — Type

parses:

a
a::B
::B
a = c
a::B = c

ExprParsers.Assignment — Type

parses:

left = right

ExprParsers.Assignment_Parsed — Type

parses:

left = right

ExprParsers.Block — Type

parses standard blocks of code or Vectors of Base.Expr

ExprParsers.Call — Type

parses:

a()
a(b, c)
a{b, c}(d, e)

ExprParsers.Call_Parsed — Type

parses:

a()
a(b, c)
a{b, c}(d, e)

ExprParsers.Expr — Type

parses:

Base.Expr(head, args...)

ExprParsers.ExprParsed — Type

ExprParsed(parser::ExprParserWithParsed)::ExprParsed

Maps Parser to respective Parsed type, and is also abstract super type of all Parsed types.

Example

EP.ExprParsed(EP.Assignment) == EP.Assignment_Parsed

ExprParsers.ExprParser — Type

All parsers in the ExprParsers package inherit from this type.

ExprParsers.ExprParserWithParsed — Type

This is mainly for internal usage. Please use @exprparser instead for the public interface.

Subtype of ExprParser which indicates that this parser actually constructs a ExprParsed object when calling parse_expr. The resulting ExprParsed object is a struct with identical fields like the parser, where then the parsed values will be stored.

ExprParsed(parser::ExprParserWithParsed) will return the corresponding ExprParsed type.

ExprParsers.Expr_Parsed — Type

parses:

Base.Expr(head, args...)

ExprParsers.Function — Type

parses:

function a(b, c) where B
  d
end
a(b, c) = d

ExprParsers.Function_Parsed — Type

parses:

function a(b, c) where B
  d
end
a(b, c) = d

ExprParsers.Indexed — Type

adds a Mapping Layer to a possibly nested Parser with which you can refer into a deep nested subparser by name

use like

EP.Indexed() do dict
  EP.Expr(quote
    a = $(dict[:a] = EP.Isa(Int))
    b = $(dict[:b] = EP.anysymbol)
  end)
end

ExprParsers.Macro — Type

parses:

@macroname arg1 arg2 ...

ExprParsers.Macro_Parsed — Type

parses:

@macroname arg1 arg2 ...

ExprParsers.Named — Type

identify Parser by Tag

Named Parsers can be easily identified by Type. The Nametag is also passed to the parsed value.

ExprParsers.NestedDot — Type

parses:

a.b
fun(T{:hi}).b.c.d.e.f

ExprParsers.NestedDot_Parsed — Type

parses:

a.b
fun(T{:hi}).b.c.d.e.f

ExprParsers.Reference — Type

parses:

a
a{b, c}

ExprParsers.Reference_Parsed — Type

parses:

a
a{b, c}

ExprParsers.Signature — Type

parses:

a(b, c::Any)
a(b::B, c) where B
(::Any, c::C) where {C <: Number}

ExprParsers.Signature_Parsed — Type

parses:

a(b, c::Any)
a(b::B, c) where B
(::Any, c::C) where {C <: Number}

ExprParsers.Type — Type

parses:

a
a{b, c}
a{d} where d

ExprParsers.TypeAnnotation — Type

parses:

a
a::B
::B

ExprParsers.TypeAnnotation_Parsed — Type

parses:

a
a::B
::B

ExprParsers.TypeRange — Type

parses:

TypeVar >: LowerBound
TypeVar <: UpperBound
LowerBound <: TypeVar <: UpperBound

Note: Construct with typevar = anysymbol to guarantee that only plain symbols can be used as type variable

ExprParsers.TypeRange_Parsed — Type

parses:

TypeVar >: LowerBound
TypeVar <: UpperBound
LowerBound <: TypeVar <: UpperBound

Note: Construct with typevar = anysymbol to guarantee that only plain symbols can be used as type variable

ExprParsers.Type_Parsed — Type

parses:

a
a{b, c}
a{d} where d

ExprParsers._find_object — Method

return list of keys/symbols to access a reference or nothing if no is found

ExprParsers._get_struct_name — Method

returns name of struct definition as symbol

ExprParsers._parsed_struct_from_parser_struct — Method

makes it mutable

adds SUFFIX to struct name

deletes all default values

ExprParsers.parse_expr — Method

parse_expr(parser, value)

Match parser against a value, will throw ParseError if the parser does not match.

Defaults to comparing with ==, if matches, will return the value. Parsers will be called instead and return their parsed value.

ExprParsers.parse_expr — Method

All ExprParserWithParsed have a common parse_expr method, namely that all struct fields are given directly as keyword arguments.

ExprParsers.to_expr — Method

to_expr(parsed)

Converts parsed information back to Expr.

Defaults to returning same value, however if something knows about how it can be translated back, just overload the function.

ExprParsers.@passert — Macro

@passert cond [text]

Throw an ParseError if cond is false. Preferred syntax for writing assertions. Message text is optionally displayed upon assertion failure.

Examples

julia> @passert iseven(3) "3 is an odd number!"
ERROR: ParseError: 3 is an odd number!
julia> @passert isodd(3) "What even are numbers?"

Adapted from Base.@assert